Revise \eta_T and \eta_E notation
In many places subscripts are used for variant, eg I_v, f_v, etc... however \eta_T doesn't refer to parameter for variant T (which exists in places like Figure 1) it instead refers to "transmission" parameter. Similarly, \eta_E refers to "escape" parameter and not parameter for variant E.

There were places in the text with \eta_T^escape and it was endlessly confusing to me to figure out that ^escape is the variant and _T is the transmission parameter.

I addressed this by splitting \eta_T and \eta_E into separate Greek letters. I kept \eta for "escape" and gave it the shortcut \varEscape, and I moved "transmission" to \rho with the short-cut \varTransmission. This was a bit arbitrary. I chose a letter that hadn't been used before and that had the same energy as \eta. I tried \kappa and \alpha and wasn't as happy with either of them as \rho. I'm not tied to using \rho specifically however. I just want distinct letters for these two parameters.